Enhance test configuration & add Flatfile Article
Call `composer install` after applying this commit.

The flatfile Article is not an Eloquent model. It is represented by flatfiles. It can be loaded from and serialized to flatfiles.
Changing its path causes the flatfile to be moved. Changes are applied only after calling `$article->save()`.

Article should still be refactored for cleaner and easier to read code.

Also added the following helper methods (mostly for Unit Tests, but might be used for manual tests in artisan tinker as well):

* fake_hierarchical_slug
* make_fake_article
* create_fake_article

Configured default Storage during tests to be in 'storage/framework/testing/disks/testing' to prevent collisions of real storage with test storage.

Configured DB for testing to be sqlite. This makes the project setup easier and let's you dive in just after having cloned the project without the need to provide mysql db.
It also allows us to maintain only one MySQL DB. Sometimes this causes problems, since sqlite is not always compatible with mysql. If this happens, consider switching
to mysql for testing.

The new sqlite testing configuration requires you to run `touch database/database.sqlite` once.