Handle frontmatter edgecases well
This commit fixes the behavior of the ArticleFileLoader when dealing with:

* an empty frontmatter section
* syntactically wrong frontmatter
* unexpected frontmatter (that would not result in an assoc array)

It makes sure that unexpected frontmatter will log a warning but still provides access to the markdown.

I also configured the log path during testing to be separate from the normal log path, in order to prevent logfile pollution.

The new timacdonald/log-fake dev dependency enables us to assert logging in tests.